+++ R/with_parameters.R
@@ -27,9 +27,10 @@
 #' data frame, pass it in the `.cases` argument.
 #'
 #' One parameter is noteworthy. If the user passes a character vector as
-#' `test_name`, each instance is combined with `desc_stub` to create the
+#' `.test_name`, each instance is combined with `desc_stub` to create the
 #' completed test name. Similarly, the named argument from `cases()` is combined
-#' with `desc_stub` to create the parameterized test names.
+#' with `desc_stub` to create the parameterized test names. When names aren't
+#' provided, they will be automatically generated using the test data.
 #'
 #' @param desc_stub A string scalar. Used in creating the names of the
 #'   parameterized tests.
@@ -38,7 +39,8 @@
 #'   same length.
 #' @param .cases A data frame where each row contains test parameters.
 #' @param .test_name An alternative way for providing test names. If provided,
-#'   the name will be appended to the stub description in `desc_stub`.
+#'   the name will be appended to the stub description in `desc_stub`. If not
+#'   provided, test names will be automatically generated.
 #' @examples
 #' with_parameters_test_that("trigonometric functions match identities:",
 #'   {
@@ -62,7 +64,7 @@
 #'   )
 #' )
 #'
-#' # Or, pass a dataframe of cases, perhaps using a helper function
+#' # Or, pass a data frame of cases, perhaps using a helper function
 #' make_cases <- function() {
 #'   tibble::tribble(
 #'     ~.test_name, ~expr, ~numeric_value,
@@ -85,13 +87,13 @@ with_parameters_test_that <- function(desc_stub,
                                       code,
                                       ...,
                                       .cases = NULL,
-                                      .test_name = "") {
-  if (!is.null(.cases)) {
-    all_pars <- .cases
-  } else {
+                                      .test_name = NULL) {
+  if (is.null(.cases)) {
     pars <- tibble::tibble(...)
     possibly_add_column <- purrr::possibly(tibble::add_column, otherwise = pars)
     all_pars <- possibly_add_column(pars, .test_name = .test_name)
+  } else {
+    all_pars <- .cases
   }
   # TODO: drop this once downstream users upgrade their version of patrick.
   if ("test_name" %in% names(all_pars)) {
@@ -109,9 +111,28 @@ with_parameters_test_that <- function(desc_stub,
       test_name = NULL
     )
   }
+  if (!".test_name" %in% names(all_pars)) {
+    all_pars$.test_name <- build_test_names(all_pars)
+  }
   captured <- rlang::enquo(code)
   purrr::pmap(all_pars, build_and_run_test, desc = desc_stub, code = captured)
   invisible(TRUE)
+}
+
+#' Generate test names from cases, if none are provided.
+#'
+#' @param all_cases A tibble containing test cases.
+#' @return A character vector, whose length matches the number of rows in
+#'   `all_cases`.
+#' @noRd
+build_test_names <- function(all_cases) {
+  case_names <- names(all_cases)
+  build_label <- function(...) {
+    row <- format(list(...))
+    labels <- sprintf("%s=%s", case_names, row)
+    paste(labels, collapse = ", ")
+  }
+  purrr::pmap_chr(all_cases, build_label)
 }
 
 build_and_run_test <- function(..., .test_name, desc, code, env) {