+++ tests/testthat/test-with_parameters.R
@@ -30,9 +30,25 @@ with_parameters_test_that(
 with_parameters_test_that(
   "Names are added",
   {
-    testthat::expect_true(.test_name == "")
+    testthat::expect_identical(.test_name, "case=TRUE")
   },
   case = TRUE
+)
+
+with_parameters_test_that(
+  "Names can be extracted from cases",
+  {
+    testthat::expect_identical(
+      .test_name,
+      "logical=FALSE, number=1, string=hello"
+    )
+  },
+  .cases = data.frame(
+    logical = FALSE,
+    number = 1,
+    string = "hello",
+    stringsAsFactors = FALSE
+  )
 )
 
 with_parameters_test_that(